refactor:  getPoductsList remove proxy and getProductsById name corr as per task
