+++ shop-service/src/functions/getProductsById/index.ts
