+++ shop-service/src/functions/getProductsList/index.ts
@@ -7,7 +7,7 @@ export default {
       http: {
         method: 'get',
         path: 'products',
-        
+        corse: true
       },
     },
   ],
